feat: pull instead of push model
