--- lib/src/main/java/eu/nitonfx/signaling/DequeueSignal.java
@@ -1,48 +1,0 @@
-package eu.nitonfx.signaling;
-
-import eu.nitonfx.signaling.api.Signal;
-
-import java.util.*;
-import java.util.function.Consumer;
-import java.util.function.Supplier;
-
-import static java.util.Set.copyOf;
-
-
-public class DequeueSignal<T> implements Subscribable, Signal<T> {
-    private final Consumer<Subscribable> readCallback;
-    private final Set<Runnable> observers = new HashSet<>();
-    private final Consumer<Supplier<Set<Runnable>>> writeCallback;
-    private T value;
-    private final StackTraceElement trace;
-
-    public DequeueSignal(Consumer<Subscribable> readCallback, Consumer<Supplier<Set<Runnable>>> writeCallback, T value) {
-        this.readCallback = readCallback;
-        this.writeCallback = writeCallback;
-        this.value = value;
-        trace = Thread.currentThread().getStackTrace()[3];
-    }
-
-    public Subscription subscribe(Runnable observer) {
-        observers.add(observer);
-        return () -> observers.remove(observer);
-    }
-
-    @Override
-    public void set(T i) {
-        if(Objects.equals(i, value)) return;
-        value = i;
-        writeCallback.accept(()->copyOf(observers));
-    }
-
-    @Override
-    public T get() {
-        readCallback.accept(this);
-        return value;
-    }
-
-    @Override
-    public T getUntracked() {
-        return value;
-    }
-}
